refactor the database code
