+++ server/services/admin/Mutation.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"github.com/kataras/go-sessions/v3"
 	"log"
-	"server/services/database"
+	"server/services/database/resolvers"
 	"server/services/types"
 	"strconv"
 )
@@ -22,10 +22,10 @@ func (r *MutationResolver) BulletinPub(input types.BulletinPubInput, ctx context
 	if input.CheckPass() {
 		return &types.BulletinPubResult{Message: "not empty error"}, nil
 	}
-	err := database.AddBulletin(input.Title, input.Content, input.Topping)
+	err := resolvers.AddBulletin(input.Title, input.Content, input.Topping)
 	if err != nil {
-		log.Println("Bulletin addition Error: ", err.Error())
-		return &types.BulletinPubResult{Message: "Bulletin addition Error!"}, nil
+		log.Println("resolvers addition Error: ", err.Error())
+		return &types.BulletinPubResult{Message: "resolvers addition Error!"}, nil
 	}
 	return &types.BulletinPubResult{Message: ""}, nil
 
@@ -42,7 +42,7 @@ func (r *MutationResolver) UserInfoUpdate(input types.UserInfoUpdateInput, ctx c
 		return &types.UserInfoUpdateResult{Message: "user information check failed"}, nil
 	}
 	if userId != nil {
-		checkResult, err := database.CheckAdminByUserId(*userId)
+		checkResult, err := resolvers.CheckAdminByUserId(*userId)
 		if err != nil {
 			return nil, err
 		}
@@ -54,7 +54,7 @@ func (r *MutationResolver) UserInfoUpdate(input types.UserInfoUpdateInput, ctx c
 		if checkResult && inputUserId == *userId && !(input.Role == "admin") {
 			return &types.UserInfoUpdateResult{Message: "downgrade not permitted"}, nil
 		}
-		err = database.UpdateUserInfo(inputUserId, input.Name, input.Role, input.Mail, input.State)
+		err = resolvers.UpdateUserInfo(inputUserId, input.Name, input.Role, input.Mail, input.State)
 		if err != nil {
 			return nil, err
 		}
@@ -75,14 +75,14 @@ func (r *MutationResolver) ChallengeMutate(input types.ChallengeMutateInput, ctx
 		return &types.ChallengeMutateResult{Message: "Challenge format not available"}, nil
 	}
 	if input.ChallengeId == "" {
-		err := database.AddChallenge(input)
+		err := resolvers.AddChallenge(input)
 		if err != nil {
 			return nil, err
 		}
 		return &types.ChallengeMutateResult{Message: "Challenge format not available"}, nil
 	}
 
-	err := database.UpdateChallenge(input)
+	err := resolvers.UpdateChallenge(input)
 	if err != nil {
 		return nil, err
 	}
