+++ server/services/database/resolvers/Common.go
@@ -1,0 +1,5 @@
+package resolvers
+
+import "server/services/database"
+
+var db = database.DataBase
