+++ server/services/database/resolvers/ReplicaAlloc.go
@@ -1,0 +1,18 @@
+package resolvers
+
+import (
+	"errors"
+	"gorm.io/gorm"
+	"server/entity"
+)
+
+func FindReplicaAllocByUserId(userId uint64) ([]entity.ReplicaAlloc, error) {
+	var replicaAllocs []entity.ReplicaAlloc
+	result := db.Where(map[string]interface{}{"UserId": userId}).Find(&replicaAllocs)
+	if errors.Is(result.Error, gorm.ErrRecordNotFound) {
+		return nil, nil
+	} else if result.Error != nil {
+		return nil, result.Error
+	}
+	return replicaAllocs, nil
+}
